refactor: ログイン方式をスクレイピング戦略から分離する
AssetScrapingStrategy に混在していたログイン関連メソッド
(isSessionValid / login) を MatsuiLoginMethod インターフェースとして
切り出した。現行のパスワード認証は PasswordLoginMethod として実装し、
FundStrategy / UsStockStrategy の重複コードを統合した。
