+++ src/modules/matsui/login-methods/login-method.ts
@@ -1,0 +1,19 @@
+import type { Page } from "playwright";
+
+/**
+ * 松井証券へのログイン方式のインターフェース
+ */
+export interface MatsuiLoginMethod {
+  /**
+   * セッションの有効性を確認する
+   * @param page PlaywrightのPageオブジェクト
+   * @returns セッションが有効かどうか
+   */
+  isSessionValid(page: Page): Promise<boolean>;
+
+  /**
+   * ログイン処理を実行する
+   * @param page PlaywrightのPageオブジェクト
+   */
+  login(page: Page): Promise<void>;
+}
