+++ src/modules/matsui/strategies/strategy-interface.ts
@@ -5,19 +5,6 @@ import type { Page } from "playwright";
  */
 export interface AssetScrapingStrategy<T> {
   /**
-   * セッションの有効性を確認する
-   * @param page PlaywrightのPageオブジェクト
-   * @returns セッションが有効かどうか
-   */
-  isSessionValid(page: Page): Promise<boolean>;
-
-  /**
-   * ログイン処理を実行する
-   * @param page PlaywrightのPageオブジェクト
-   */
-  login(page: Page): Promise<void>;
-
-  /**
    * スクレイピング対象のページを準備する（オプショナル）
    * @param page PlaywrightのPageオブジェクト
    * @returns スクレイピングに使用するPageオブジェクト
